+++ src/components/ContributionAssistantPriceFormCard.vue
@@ -16,12 +16,12 @@
       <ProductInputRow :productForm="productPriceForm" :disableInitWhenSwitchingType="true" :hideProductBarcode="false" :hideBarcodeScannerTab="true" @filled="productFormFilled = $event" />
       <PriceInputRow class="mt-0" :priceForm="productPriceForm" :product="productPriceForm.product" :hideCurrencyChoice="true" @filled="pricePriceFormFilled = $event" />
     </v-card-text>
-    <v-divider v-if="mode === 'Validation'" />
-    <v-card-text v-if="mode === 'Validation'">
+    <v-divider v-if="!hideProofDetails" />
+    <v-card-text v-if="!hideProofDetails">
       <ProofFooterRow :proof="productPriceForm.proof" :showProofChip="true" :hideProofType="true" :hideProofActions="true" :readonly="true" />
     </v-card-text>
-    <v-divider />
-    <v-card-actions>
+    <v-divider v-if="!hideActions" />
+    <v-card-actions v-if="!hideActions">
       <v-btn
         color="error"
         variant="outlined"
@@ -59,9 +59,9 @@ import constants from '../constants'
 export default {
   components: {
     ProofImageCropped: defineAsyncComponent(() => import('../components/ProofImageCropped.vue')),
-  ProductInputRow: defineAsyncComponent(() => import('../components/ProductInputRow.vue')),
-  PriceInputRow: defineAsyncComponent(() => import('../components/PriceInputRow.vue')),
-  ProofFooterRow: defineAsyncComponent(() => import('../components/ProofFooterRow.vue')),
+    ProductInputRow: defineAsyncComponent(() => import('../components/ProductInputRow.vue')),
+    PriceInputRow: defineAsyncComponent(() => import('../components/PriceInputRow.vue')),
+    ProofFooterRow: defineAsyncComponent(() => import('../components/ProofFooterRow.vue')),
   },
   props: {
     productPriceForm: {
@@ -88,6 +88,14 @@ export default {
       default: 'Contribution'  // or 'Validation'
     },
     showProductNameField: {
+      type: Boolean,
+      default: false
+    },
+    hideProofDetails: {
+      type: Boolean,
+      default: false
+    },
+    hideActions: {
       type: Boolean,
       default: false
     }