+++ src/components/ProofTypeInputRow.vue
@@ -3,7 +3,7 @@
     <v-col cols="12">
       <v-item-group v-model="proofTypeForm.type" class="d-inline" mandatory>
         <v-item v-for="pt in proofTypeList" :key="pt.key" v-slot="{ isSelected, toggle }" :value="pt.key">
-          <v-chip class="mr-1" :class="isSelected ? 'border-success' : ''" variant="outlined" @click="toggle">
+          <v-chip class="mr-1" :class="isSelected ? 'border-success' : ''" variant="outlined" density="comfortable" @click="toggle">
             <v-icon start :icon="pt.icon" />
             {{ $t('ProofCard.' + pt.value) }}
             <v-icon end :icon="isSelected ? 'mdi-checkbox-marked-circle' : 'mdi-circle-outline'" :color="isSelected ? 'green' : ''" />