+++ src/utils.js
@@ -233,6 +233,17 @@ function getLocaleOriginTags(locale) {
   return import(`./data/origins/${locale}.json`)
 }
 
+function getLocaleLabelTags(locale) {
+  return import(`./data/labels/${locale}.json`)
+}
+
+function getLocaleLabelTagName(locale, labelId) {
+  return getLocaleLabelTags(locale).then((module) => {
+    let label = module.default.find(ct => ct.id === labelId)
+    return label ? label.name : labelId
+  })
+}
+
 function getPriceTypeIcon(priceType) {
   return constants[`PRICE_TYPE_${priceType}_ICON`] || constants.PRICE_ICON
 }
@@ -444,6 +455,8 @@ export default {
   getLocaleCategoryTag,
   getLocaleCategoryTagName,
   getLocaleOriginTags,
+  getLocaleLabelTags,
+  getLocaleLabelTagName,
   getPriceTypeIcon,
   getProofTypeIcon,
   getCountryEmojiFromName,